finished refactor of file structure
