+++ src/ai/__init__.py
@@ -1,0 +1,7 @@
+# src/ai/__init__.py
+# AI model clients and related functionality
+
+from .prompts import build_sectionizer_prompt, build_generate_prompt
+from .clients.openai_client import run_generate
+
+__all__ = ["build_sectionizer_prompt", "build_generate_prompt", "run_generate"]
