+++ src/ai/clients/__init__.py
@@ -1,0 +1,6 @@
+# src/ai/clients/__init__.py
+# AI client implementations
+
+from .openai_client import run_generate
+
+__all__ = ["run_generate"]
