+++ src/ai/prompts.py
@@ -1,4 +1,4 @@
-# src/prompts.py
+# src/ai/prompts.py
 # Prompt templates for AI-powered resume sectionizing and tailoring operations
 
 # build sectionizer prompt for LLM
