+++ src/config/__init__.py
@@ -1,0 +1,6 @@
+# src/config/__init__.py
+# Configuration management
+
+from .settings import settings_manager
+
+__all__ = ["settings_manager"]
