+++ src/core/__init__.py
@@ -1,0 +1,6 @@
+# src/core/__init__.py
+# Core business logic
+
+from . import pipeline
+
+__all__ = ["pipeline"]
